Moved 'get_birthdays' functionality to classes.py. Deleted AddressBook.py, which is a file from the old version of <main> branch. Added test_birthdays.py
